Issue #11025 - Refactor create functions
